Started to add Weather Forecast with CSS,JS,HTML
CSS - removed the .weather-temperature class and .future class.  added the .weather-forecast class
html - added columns with bootstrap classes and created basic outline for weather forecast.

JS- created displayForecast function that contains an Array of weekdays. created a forEach function that will call the Array and the block of HTML code   with the variable forecastHTML